Move Vulkan objects from inner class to own class
