+++ graphics-by-opengl-j2se/src/main/java/com/nucleus/vulkan/ExtensionProperties.java
@@ -1,0 +1,16 @@
+package com.nucleus.vulkan;
+
+public class ExtensionProperties {
+    public ExtensionProperties(String name, int specVersion) {
+        this.name = name;
+        this.specVersion = specVersion;
+    }
+
+    String name;
+    int specVersion;
+
+    public String getName() {
+        return name;
+    }
+
+}
