+++ graphics-by-opengl-j2se/src/main/java/com/nucleus/vulkan/Extent3D.java
@@ -1,0 +1,11 @@
+package com.nucleus.vulkan;
+
+public class Extent3D {
+    final int[] values = new int[3];
+
+    public Extent3D(int width, int height, int depth) {
+        values[0] = width;
+        values[1] = height;
+        values[2] = depth;
+    }
+}
