+++ graphics-by-opengl-lwjgl3/src/main/java/com/nucleus/vulkan/LWJGLVulkanDeviceFeatures.java
@@ -4,8 +4,6 @@ import org.lwjgl.vulkan.VK10;
 import org.lwjgl.vulkan.VkPhysicalDevice;
 import org.lwjgl.vulkan.VkPhysicalDeviceFeatures;
 
-import com.nucleus.vulkan.VulkanWrapper.PhysicalDeviceFeatures;
-
 public class LWJGLVulkanDeviceFeatures extends PhysicalDeviceFeatures {
 
     public LWJGLVulkanDeviceFeatures(VkPhysicalDevice device, long surface) {
